feat(): redesign `Network` to `Aim`
