--- Package.resolved
@@ -1,14 +1,0 @@
-{
-  "pins" : [
-    {
-      "identity" : "rxswift",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/ReactiveX/RxSwift",
-      "state" : {
-        "revision" : "b4307ba0b6425c0ba4178e138799946c3da594f8",
-        "version" : "6.5.0"
-      }
-    }
-  ],
-  "version" : 2
-}
