+++ Sources/Aim/Request/NoneRequest.swift
