--- Sources/Network/Response.swift
@@ -1,10 +1,0 @@
-//
-//  Response.swift
-//  
-//
-//  Created by JSilver on 2023/02/13.
-//
-
-import Foundation
-
-public typealias Response = Result<(Data, URLResponse), any Error>
